requirements.txt
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
fastapi==0.65.1
h11==0.12.0
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.13.4
